+++ src/NLog/Config/XmlLoggingConfiguration.cs
@@ -59,6 +59,8 @@ namespace NLog.Config
     /// <summary>
     /// A class for configuring NLog through an XML configuration file 
     /// (App.config style or App.nlog style).
+    /// 
+    /// Parsing of the XML file is also implemented in this class.
     /// </summary>
     ///<remarks>This class is thread-safe.<c>.ToList()</c> is used for that purpose.</remarks>
     public class XmlLoggingConfiguration : LoggingConfiguration
@@ -99,7 +101,7 @@ namespace NLog.Config
         /// Initializes a new instance of the <see cref="XmlLoggingConfiguration" /> class.
         /// </summary>
         /// <param name="fileName">Configuration file to be read.</param>
-        /// <param name="ignoreErrors">Ignore any errors during configuration.</param>
+        /// <param name="logFactory">The <see cref="LogFactory" /> to which to apply any applicable configuration values.</param>
         public XmlLoggingConfiguration(string fileName, LogFactory logFactory)
             : this(fileName, false, logFactory)
         { }
@@ -132,7 +134,6 @@ namespace NLog.Config
         /// <summary>
         /// Initializes a new instance of the <see cref="XmlLoggingConfiguration" /> class.
         /// </summary>
-        /// <param name="reader"><see cref="XmlReader"/> containing the configuration section.</param>
         /// <param name="fileName">Name of the file that contains the element (to be used as a base for including other files).</param>
         private static XmlReader CreateFileReader(string fileName)
         {
@@ -168,7 +169,8 @@ namespace NLog.Config
         /// </summary>
         /// <param name="reader"><see cref="XmlReader"/> containing the configuration section.</param>
         /// <param name="fileName">Name of the file that contains the element (to be used as a base for including other files).</param>
-        /// <param name="ignoreErrors">Ignore any errors during configuration.</param>
+        /// <param name="logFactory">The <see cref="LogFactory" /> to which to apply any applicable configuration values.</param>
+
         public XmlLoggingConfiguration(XmlReader reader, string fileName, LogFactory logFactory)
             : this(reader, fileName, false, logFactory)
         { }
@@ -222,6 +224,8 @@ namespace NLog.Config
         /// <param name="ignoreErrors">If set to <c>true</c> errors will be ignored during file processing.</param>
         internal XmlLoggingConfiguration(XmlElement element, string fileName, bool ignoreErrors)
         {
+            logFactory = LogManager.LogFactory;
+
             using (var stringReader = new StringReader(element.OuterXml))
             {
                 XmlReader reader = XmlReader.Create(stringReader);
@@ -296,6 +300,32 @@ namespace NLog.Config
         public override LoggingConfiguration Reload()
         {
             return new XmlLoggingConfiguration(this.originalFileName);
+        }
+
+        /// <summary>
+        /// Get file paths (including filename) for the possible NLog config files. 
+        /// </summary>
+        /// <returns>The filepaths to the possible config file</returns>
+        public static IEnumerable<string> GetCandidateConfigFilePaths()
+        {
+            return LogManager.LogFactory.GetCandidateConfigFilePaths();
+        }
+
+        /// <summary>
+        /// Overwrite the paths (including filename) for the possible NLog config files.
+        /// </summary>
+        /// <param name="filePaths">The filepaths to the possible config file</param>
+        public static void SetCandidateConfigFilePaths(IEnumerable<string> filePaths)
+        {
+            LogManager.LogFactory.SetCandidateConfigFilePaths(filePaths);
+        }
+
+        /// <summary>
+        /// Clear the candidate file paths and return to the defaults.
+        /// </summary>
+        public static void ResetCandidateConfigFilePath()
+        {
+            LogManager.LogFactory.ResetCandidateConfigFilePath();
         }
 
         #endregion
@@ -405,7 +435,7 @@ namespace NLog.Config
 
                 if (!ignoreErrors)
                 {
-                    if (exception.MustBeRethrown())
+                    if (configurationException.MustBeRethrown())
                     {
                         throw configurationException;
                     }
@@ -914,7 +944,26 @@ namespace NLog.Config
                 string type = StripOptionalNamespacePrefix(addElement.GetOptionalAttribute("type", null));
                 if (type != null)
                 {
+                    try
+                    {
                     this.ConfigurationItemFactory.RegisterType(Type.GetType(type, true), prefix);
+                }
+                    catch (Exception exception)
+                    {
+                        if (exception.MustBeRethrownImmediately())
+                        {
+                            throw;
+                        }
+
+                        InternalLogger.Error(exception, "Error loading extensions.");
+                        NLogConfigurationException configException =
+                            new NLogConfigurationException("Error loading extensions: " + type, exception);
+
+                        if (configException.MustBeRethrown())
+                        {
+                            throw configException;
+                        }
+                    }
                 }
 
 #if !UWP10
@@ -925,6 +974,7 @@ namespace NLog.Config
                     {
                         var asm = AssemblyHelpers.LoadFromPath(assemblyFile);
                         this.ConfigurationItemFactory.RegisterItemsFromAssembly(asm, prefix);
+
                     }
                     catch (Exception exception)
                     {
@@ -934,10 +984,12 @@ namespace NLog.Config
                         }
 
                         InternalLogger.Error(exception, "Error loading extensions.");
+                        NLogConfigurationException configException =
+                            new NLogConfigurationException("Error loading extensions: " + assemblyFile, exception);
 
-                        if (exception.MustBeRethrown())
+                        if (configException.MustBeRethrown())
                         {
-                            throw new NLogConfigurationException("Error loading extensions: " + assemblyFile, exception);
+                            throw configException;
                         }
                     }
 
@@ -956,17 +1008,18 @@ namespace NLog.Config
                     }
                     catch (Exception exception)
                     {
-
                         if (exception.MustBeRethrownImmediately())
                         {
                             throw;
                         }
 
                         InternalLogger.Error(exception, "Error loading extensions.");
+                        NLogConfigurationException configException =
+                            new NLogConfigurationException("Error loading extensions: " + assemblyName, exception);
 
-                        if (exception.MustBeRethrown())
+                        if (configException.MustBeRethrown())
                         {
-                            throw new NLogConfigurationException("Error loading extensions: " + assemblyName, exception);
+                            throw configException;
                         }
                     }
                 }
@@ -1082,7 +1135,12 @@ namespace NLog.Config
             if (elementType != null)
             {
                 IList propertyValue = (IList)propInfo.GetValue(o, null);
-                object arrayItem = FactoryHelper.CreateInstance(elementType);
+
+                object arrayItem = TryCreateLayoutInstance(element, elementType);
+                // arrayItem is not a layout
+                if (arrayItem == null)
+                    arrayItem = FactoryHelper.CreateInstance(elementType);
+
                 this.ConfigureObjectFromAttributes(arrayItem, element, true);
                 this.ConfigureObjectFromElement(arrayItem, element);
                 propertyValue.Add(arrayItem);
@@ -1117,23 +1175,17 @@ namespace NLog.Config
             // if property exists
             if (PropertyHelper.TryGetPropertyInfo(o, name, out targetPropertyInfo))
             {
-                // and is a Layout
-                if (typeof(Layout).IsAssignableFrom(targetPropertyInfo.PropertyType))
-                {
-                    string layoutTypeName = StripOptionalNamespacePrefix(layoutElement.GetOptionalAttribute("type", null));
+                Layout layout = TryCreateLayoutInstance(layoutElement, targetPropertyInfo.PropertyType);
 
-                    // and 'type' attribute has been specified
-                    if (layoutTypeName != null)
+                // and is a Layout and 'type' attribute has been specified
+                if (layout != null)
                     {
-                        // configure it from current element
-                        Layout layout = this.ConfigurationItemFactory.Layouts.CreateInstance(this.ExpandSimpleVariables(layoutTypeName));
                         this.ConfigureObjectFromAttributes(layout, layoutElement, true);
                         this.ConfigureObjectFromElement(layout, layoutElement);
                         targetPropertyInfo.SetValue(o, layout, null);
                         return true;
                     }
                 }
-            }
 
             return false;
         }
@@ -1193,6 +1245,21 @@ namespace NLog.Config
 
             InternalLogger.Debug("Wrapping target '{0}' with '{1}' and renaming to '{2}", wrapperTargetInstance.Name, wrapperTargetInstance.GetType().Name, t.Name);
             return wrapperTargetInstance;
+        }
+
+        private Layout TryCreateLayoutInstance(NLogXmlElement element, Type type)
+        {
+            // Check if it is a Layout
+            if (!typeof(Layout).IsAssignableFrom(type))
+                return null;
+
+            string layoutTypeName = StripOptionalNamespacePrefix(element.GetOptionalAttribute("type", null));
+
+            // Check if the 'type' attribute has been specified
+            if (layoutTypeName == null)
+                return null;
+
+            return this.ConfigurationItemFactory.Layouts.CreateInstance(this.ExpandSimpleVariables(layoutTypeName));
         }
 
         /// <summary>