+++ src/NLog/LogFactory.cs
@@ -84,6 +84,12 @@ namespace NLog
         private readonly LoggerCache loggerCache = new LoggerCache();
 
         /// <summary>
+        /// Overwrite possible file paths (including filename) for possible NLog config files. 
+        /// When this property is <c>null</c>, the default file paths (<see cref="GetCandidateConfigFilePaths"/> are used.
+        /// </summary>
+        private List<string> candidateConfigFilePaths;
+
+        /// <summary>
         /// Occurs when logging <see cref="Configuration" /> changes.
         /// </summary>
         public event EventHandler<LoggingConfigurationChangedEventArgs> ConfigurationChanged;
@@ -167,7 +173,7 @@ namespace NLog
                     if (this.configLoaded)
                         return this.config;
 
-#if !SILVERLIGHT  && !__IOS__ && !__ANDROID__ && !UWP10
+#if !SILVERLIGHT && !__IOS__ && !__ANDROID__ && !UWP10
                     if (this.config == null)
                     {
                         // Try to load default configuration.
@@ -177,7 +183,8 @@ namespace NLog
                     // Retest the condition as we might have loaded a config.
                     if (this.config == null)
                     {
-                        foreach (string configFile in GetCandidateConfigFileNames())
+                        var configFileNames = GetCandidateConfigFilePaths();
+                        foreach (string configFile in configFileNames)
                         {
 #if SILVERLIGHT
                             Uri configFileUri = new Uri(configFile, UriKind.Relative);
@@ -418,7 +425,7 @@ namespace NLog
 
 #if !UWP10
         /// <summary>
-        /// Gets a custom logger with the name of the current class. Use <paramref name="loggerType"/> to pass the type of the needed Logger.
+        /// Gets a custom logger with the name of the current class. Use <typeparamref name="T"/> to pass the type of the needed Logger.
         /// </summary>
         /// <returns>The logger with type <typeparamref name="T"/>.</returns>
         /// <typeparam name="T">Type of the logger</typeparam>
@@ -442,7 +449,8 @@ namespace NLog
         /// <param name="loggerType">The type of the logger to create. The type must inherit from 
         /// NLog.Logger.</param>
         /// <returns>The logger of type <paramref name="loggerType"/>.</returns>
-        /// <remarks>This is a slow-running method. Make sure you are not calling this method in a 
+        /// <remarks>This is a slow-running method.
+        /// Make sure you're not doing this in a loop.</remarks>
         [MethodImpl(MethodImplOptions.NoInlining)]
         public Logger GetCurrentClassLogger(Type loggerType)
         {
@@ -468,18 +476,17 @@ namespace NLog
         }
 
         /// <summary>
-        /// Gets the custom named logger. Use <paramref name="loggerType"/> to pass the type of the needed Logger.
+        /// Gets the custom named logger. Use <typeparamref name="T"/>  to pass the type of the needed Logger.
         /// </summary>
         /// <param name="name">Name of the logger.</param>
-        /// <param name="loggerType">The type of the logger to create. The type must inherit from <see cref="Logger"/>.	</param>       
-        /// <typeparam name="T">Type of the logger</typeparam>
+        /// <typeparam name="T">The type of the logger to create. The type must inherit from <see cref="Logger"/>.	</typeparam>       
         /// <returns>The logger reference with type <typeparamref name="T"/>. Multiple calls to <c>GetLogger</c> with the same argument 
         /// are not guaranteed to return the same logger reference.</returns>
         public T GetLogger<T>(string name) where T : Logger
         {
             return (T)this.GetLogger(new LoggerCacheKey(name, typeof(T)));
         }
-        /// same argument aren't guaranteed to return the same logger reference.</returns>
+
         /// <summary>
         /// Gets the specified named logger.  Use <paramref name="loggerType"/> to pass the type of the needed Logger.
         /// </summary>
@@ -842,6 +849,7 @@ namespace NLog
             }
 
             InternalLogger.Debug("Targets for {0} by level:", name);
+
             for (int i = 0; i <= LogLevel.MaxLevel.Ordinal; ++i)
             {
                 StringBuilder sb = new StringBuilder();
@@ -884,7 +892,45 @@ namespace NLog
 #endif
         }
 
-        private static IEnumerable<string> GetCandidateConfigFileNames()
+        /// <summary>
+        /// Get file paths (including filename) for the possible NLog config files. 
+        /// </summary>
+        /// <returns>The filepaths to the possible config file</returns>
+        public IEnumerable<string> GetCandidateConfigFilePaths()
+        {
+            if (candidateConfigFilePaths != null)
+            {
+                return candidateConfigFilePaths.AsReadOnly();
+            }
+
+            return GetDefaultCandidateConfigFilePaths();
+        }
+
+        /// <summary>
+        /// Overwrite the paths (including filename) for the possible NLog config files.
+        /// </summary>
+        /// <param name="filePaths">The filepaths to the possible config file</param>
+        public void SetCandidateConfigFilePaths(IEnumerable<string> filePaths)
+        {
+            candidateConfigFilePaths = new List<string>();
+
+            if (filePaths != null)
+            {
+                candidateConfigFilePaths.AddRange(filePaths);
+            }
+        }
+        /// <summary>
+        /// Clear the candidate file paths and return to the defaults.
+        /// </summary>
+        public void ResetCandidateConfigFilePath()
+        {
+            candidateConfigFilePaths = null;
+        }
+
+        /// <summary>
+        /// Get default file paths (including filename) for possible NLog config files. 
+        /// </summary>
+        private static IEnumerable<string> GetDefaultCandidateConfigFilePaths()
         {
 #if SILVERLIGHT || __ANDROID__ || __IOS__ || UWP10
             //try.nlog.config is ios/android/silverlight